Change the whole directory, urder all the file
