--- public/home/signup.html
@@ -1,182 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <!-- Design by foolishdeveloper.com -->
-    <title>Signup Page</title>
-
-    <link rel="preconnect" href="https://fonts.gstatic.com">
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.4/css/all.min.css">
-    <link href="https://fonts.googleapis.com/css2?family=Poppins:wght@300;500;600&display=swap" rel="stylesheet">
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.6.2/dist/css/bootstrap.min.css"
-        integrity="sha384-xOolHFLEh07PJGoPkLv1IbcEPTNtaed2xpHsD9ESMhqIYd0nLMwNLD69Npy4HI+N" crossorigin="anonymous">
-    <!--Stylesheet-->
-    <style media="screen">
-        *,
-        *:before,
-        *:after {
-            padding: 0;
-            margin: 0;
-            box-sizing: border-box;
-        }
-
-        body {
-            background-color: rgba(255, 255, 255, 255);
-        }
-
-        .background {
-            width: 430px;
-            height: 520px;
-            position: absolute;
-            transform: translate(-50%, -50%);
-            left: 50%;
-            top: 50%;
-        }
-
-        .background .shape {
-            height: 200px;
-            width: 200px;
-            position: absolute;
-            border-radius: 50%;
-        }
-
-
-        .fomrs{
-            width: 100%;
-            height: 100%;
-            display: flex;
-            justify-content: center;
-            padding-top: 40px;
-        }
-        form {
-            height: 520px;
-            width: 400px;
-            top: 58%;
-            left: 50%;
-            border-radius: 10px;
-            backdrop-filter: blur(10px);
-            border: 2px solid rgba(255, 255, 255, 0.1);
-            box-shadow: 0 0 40px rgba(8, 7, 16, 0.6);
-            padding: 50px 35px;
-            background-color: blue;
-            opacity: 0.6;
-        }
-
-        form * {
-            font-family: 'Poppins', sans-serif;
-            color: #ffffff;
-            letter-spacing: 0.5px;
-            outline: none;
-            border: none;
-
-        }
-
-        form h3 {
-            font-size: 32px;
-            font-weight: 500;
-            line-height: 42px;
-            text-align: center;
-
-        }
-
-        label {
-            display: block;
-            margin-top: 30px;
-            font-size: 16px;
-            font-weight: 500;
-        }
-
-        input {
-            display: block;
-            height: 40px;
-            width: 100%;
-            background-color: rgba(255, 255, 255, 0.07);
-            border-radius: 3px;
-            padding: 0 10px;
-            margin-top: 2px;
-            font-size: 14px;
-            font-weight: 300;
-        }
-
-        ::placeholder {
-            color: #e5e5e5;
-        }
-
-        button {
-            margin-top: 14px;
-            width: 100%;
-            background-color: #ffffff;
-            color: #080710;
-            padding: 15px 0;
-            font-size: 18px;
-            font-weight: 600;
-            border-radius: 5px;
-            cursor: pointer;
-        }
-
-        .social {
-            margin-top: 30px;
-            display: flex;
-        }
-
-        .social div {
-            background: red;
-            width: 150px;
-            border-radius: 3px;
-            padding: 5px 10px 10px 5px;
-            background-color: rgba(255, 255, 255, 0.27);
-            color: rgba(255, 255, 255, 0.27);
-            text-align: center;
-        }
-
-        .social div:hover {
-            background-color: rgba(255, 255, 255, 0.47);
-        }
-
-        .social .fb {
-            margin-left: 25px;
-        }
-
-        .social i {
-            margin-right: 4px;
-        }
-
-        .header {
-            height: 110px;
-            font-size: 35px;
-        }
-    </style>
-
-</head>
-
-<body>
-    <div class="row col-12 d-flex align-items-center p-0 m-0 p-lg-4 bg-primary header">
-        <div style="color: white;">IT Department</div>
-    </div>
-
-    <div class="fomrs">
-        <form id="signupForm" method="post" onsubmit="return validatesignupForm()">
-
-            <h3>Signup Here</h3>
-
-            <label for="username">Username</label>
-            <input type="text" placeholder="Username" name="username" id="username">
-
-            <label for="email">Email</label>
-            <input type="email" placeholder="Email or Phone" name="email" id="email">
-
-
-
-            <label for="password">Password</label>
-            <input type="password" placeholder="Password" id="password" name="password">
-
-
-            <button type="submit">Signup</button>
-            <a href="login.html">i have already account</a>
-        </form>
-    </div>
-
-    <script src="validate.js"></script>
-</body>
-
-</html>
