--- public/home/test.txt
@@ -1,1 +1,0 @@
-Khizar gaandu hai
